Update api/irradiance/direct.py
This has been left somewhat behind as the focus was on the time series version. And likely it'll be removed. For the sake of a cross-check, still kept to compare if a single timestamp fed to the time series version produces the same as this version for the direct irradiance.