Fix: write_irradiance_csv() : exclude FINGERPRINT_COLUMN_NAME
It is unnecessary to repeat in each and every row the fingerprint. Hence, remove it from the dictionary written as a CSV file. Instead, make fingerprint, if given, part of the output CSV filename.

Idea : perhaps give the user an option for including, or not, the fingerprint as part of the output CSV filename ?