refactor!: Update algorithms/jenco/solar_incidence.py -- Attempt to Understand It!
The solar incidence angle definition by Hofierka (2002), based on the definition given by Jenco (1992) is eithe : a) not implemented correctly (i.e. it misses some trick to keep the derived angles in the correct range) or b) is not working right for our use-case.
Looking at the definitions of it, over and over again, it is clear that the definition by Hofierka measures azimuth angles (both the solar azimuth angle and the suface orientation angle) from East, or else based on East as a reference.
As much as I forgot the integrate the conversion function (from North-based to East-based convention) in to the time series solar incidence angle (for the input parameter 'surface_orientation' for example in calculate_solar_incidence_time_series_jenco()) which I have previously implemented and used for my first trials on the solar positioning scheme, it still does not work correctly!
This results essentially in giving a wrong orientation angle to the solar irradiance calculation functions, especially the direct inclined component.
To be explained further in a relevant issue in the project source code repository!