refactor!: Enable performance analysis via the power/broadband CLI
Add analysis (and nomenclature, yet to use!) and restructure the output section.
This implies the definition of Typer relevant bits for analysis and nomenclature.
Important is the use of a callback function for the verbose option : As it stands, the analysis of photovoltaic performance is only possible along with verbose = 7 ! Hence, the new callback function callback_analysis_of_performance() does this : if analysis is requested, it sets verbose to 7.

# Conflicts:
#	pvgisprototype/cli/power/broadband.py