refactor(cli): Rework the timestamping engine
See the callback function callback_generate_datetime_series(). In short, timestamps can be generated by 3 out of the 4 relevant parameters start_time, end_time, frequency and periods exclusively defined by the user. This is the default behavior. However, if a space-time dataset or datasets are fed by the user via the "typical" options available in most command line tools, such as the --global-irradiance-component, then the user defined parameters will be used to _filter_ the timestamps retrieved from the input space-time data. This is so in order to respect the "data".