refactor!: Refactor input for timestamps
At least, and to start with, in photovoltaic_power_output_series() : make use of str(Timestamp.now()) instead of str(now_utc_datetimezone()).
From this, fix a few functions "downstream" to properly handle a single Timestamp.
For handle the case manually when start and end time are not requested :

    elif start_time is None and end_time is None:
        timestamps = Timestamp.now(tz='UTC')

Hence, refactor bits in :

- analyse_photovoltaic_performance()
- report_photovoltaic_performance()
- calculate_photovoltaic_power_output_series()
- determine_frequency() -- This one is very important as it will set the frequenct to 'Single' !
- print_change_percentages_panel()
- convert_series_to_sparkline()

Refers to #220