refactor!: Update the single timestamp parser function
For _now_ let the parser also return a None ! This is to get going. The "right" way still to-do is to fully embrace NaTType and use wherever applicable. This will harmonise and deepen the dependency to and use of Pandas' timestamping engine but hopefull for a good reason.

Refers to #220