refactor!: Modularise retrieval of time series for api/power/broadband.py
Attention : this commit is expected to break the Web-to-Core-API !

Move retrieval of time series data out of the core API function power/broadband.py. The reason is to abstract away the "generation" of the time series that feed the calculate_photovoltaic_power_output_series() function. It may be a signle value (floar or int) or a series of values (either hand-fed through an interface or retrieved from time series data files) which, indeed, need to be fed as a NumPy Array.

Refers to #244