feature: Add solar event times in position commands event, overview
This is a larger commit that adds the calculation and output of solar event types and times.-
Changes touch upon the complete processing chain of PVGIS' internals : from the core data model definitions for EventHourAngle and EventTime, the algorithm behind event times (see : algorithms/noaa/event_time.py), the relevant API and CLI functions as well as printing the output.
Refer to the committed changes for details.

Refers to #313