fix: Suggest installation of functiontrace, error out if not under Linux/Mac
This commit introduces checks regarding the functiontrace package which is released under a free non commercial license.
Our software will suggest to let the user install it first as well as warn if the underlying OS is not Linux or MacOS.
In this way, we can still distribute our software under a strong copyleft license while, at the same time, still have built-in support for functiontrace.

Kudos to Adeline-Raluca Toader (European Commission, JRC.A4) for her support in all things about licensing.

Refers #317