refactor!: rename reflectivity to reflected as appropriate
Reflected is the energy loss due to the reflectivity.  And this is what the time series in question here are meant to be : loss of irradiance, e.g. : 'direct_inclined_reflected' and NOT 'direct_inclined_reflectivity'.  Reflectivity is then the property of a surface (rather some ratio).  The quantity asked here is the amount of irradiance loss, in other words the absolute energy loss after the effect of the reflectivity.