refactor!: add angle_output_units to factory functions
This addition of the `angle_output_units` parameter makes building (tabular) output from core API functions (even) easier. The intention is the ever-wanting to reduce the size of the codebase, avoid duplicate code and make it easier to maintain.
Adding the parameter in question in context factory functions : parse_fields(), populate_context(), build_output().

Right after adding "it", use it too in the many solar positioning functions and not only.